+++ app/src/Games/index.js
@@ -1,5 +1,6 @@
 import * as React from "react";
 
+import Search from "../Search";
 import useApi from "../auth/useApi";
 
 import styles from "./styles.module.scss";
@@ -33,16 +34,12 @@ const initGames = [
 
 const Games = () => {
   const [games, setGames] = React.useState([]);
-  const [newGame, setNewGame] = React.useState([]);
   const { loading, apiClient } = useApi();
 
   const loadGames = React.useCallback(
     async () => setGames(await apiClient.getGames()),
     [apiClient],
   );
-  // const addGame = (game) => apiClient.addGame(game).then(loadGames);
-  const findGame = (name) => apiClient.findGame(name).then(setNewGame);
-  console.log(newGame);
   React.useEffect(() => {
     !loading && loadGames();
   }, [loading, loadGames]);
@@ -50,8 +47,7 @@ const Games = () => {
   return loading ? null : (
     <section>
       <GameList {...{ games }} />
-      <FindGame {...{ findGame }} />
-      <pre>{JSON.stringify(newGame)}</pre>
+      <Search />
     </section>
   );
 };
@@ -67,37 +63,37 @@ const GameList = ({ games }) => (
   </ul>
 );
 
-const FindGame = () => {
-  const { apiClient } = useApi();
-  const [name, setName] = React.useState("");
-  const [games, setGames] = React.useState([]);
-  const canAdd = name !== "";
+// const FindGames = () => {
+//   const { apiClient } = useApi();
+//   const [name, setName] = React.useState("");
+//   const [games, setGames] = React.useState([]);
+//   const canAdd = name !== "";
 
-  const onSubmit = (e) => {
-    e.preventDefault();
-    if (canAdd) {
-      apiClient.findGame(name).then((games) => setGames(games));
-      setName("");
-    }
-  };
+//   const onSubmit = (e) => {
+//     e.preventDefault();
+//     if (canAdd) {
+//       apiClient.findGames(name).then((games) => setGames(games));
+//       setName("");
+//     }
+//   };
 
-  return (
-    <>
-      <form {...{ onSubmit }}>
-        <label>
-          New game:{" "}
-          <input
-            onChange={(e) => setName(e.currentTarget.value)}
-            value={name}
-          />
-        </label>
-        <button disabled={!canAdd} className={styles.button}>
-          Add
-        </button>
-      </form>
-      <pre>{JSON.stringify(games, null, 2)}</pre>
-    </>
-  );
-};
+//   return (
+//     <>
+//       <form {...{ onSubmit }}>
+//         <label>
+//           New game:{" "}
+//           <input
+//             onChange={(e) => setName(e.currentTarget.value)}
+//             value={name}
+//           />
+//         </label>
+//         <button disabled={!canAdd} className={styles.button}>
+//           Add
+//         </button>
+//       </form>
+//       <pre>{JSON.stringify(games, null, 2)}</pre>
+//     </>
+//   );
+// };
 
 export default Games;
