+++ app/src/auth/useApi.js
@@ -5,7 +5,7 @@ import useAuth0 from "./useAuth0";
 const makeApi = (accessToken) => {
   const actions = {
     getGames: () => _get("/api/users/games"),
-    findGame: (name) => _get("/api/games", { name }),
+    findGames: (name) => _get("/api/games", { name }),
     addGame: (game) => _post("/api/users/games", { game }),
     addOrUpdateUser: (user) => _post("/api/users", { user }),
   };
